# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Business Platform API",
        "description": "Authentication & RBAC endpoints for the AI Business Platform. See ARCHITECTURE.md, API.md, and BACKEND.md in the repository root for the full system design.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Inventory / Products"
                ],
                "summary": "List products",
                "operationId": "4b15acb9dba7b4d577588e71f53f93be",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products returned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory / Products"
                ],
                "summary": "Create a product (also provisions its stock record)",
                "operationId": "049e36dafb20c3c4ebfdc7357d5abd03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "name",
                                    "unit_price",
                                    "cost_price"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "WIDGET-001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Widget"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19.99
                                    },
                                    "cost_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.5
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Inventory / Products"
                ],
                "summary": "Get a single product",
                "operationId": "1a5cb601b3dc74723d46a9fce673a227",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product returned"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory / Products"
                ],
                "summary": "Delete a product",
                "operationId": "986b7934e881e252b5f1d48c852bc250",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "403": {
                        "description": "Missing products.manage permission"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Inventory / Products"
                ],
                "summary": "Update a product",
                "operationId": "e7116df51d555439acfddc6b36b75926",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "name",
                                    "unit_price",
                                    "cost_price"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cost_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/knowledge-base/ask": {
            "post": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Ask a question answered from the knowledge base, with citations",
                "operationId": "34c5755a2d74b1bb0b607959dc27c0d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string"
                                    },
                                    "top_k": {
                                        "description": "Defaults to config(knowledge_base.top_k)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer with citations returned"
                    },
                    "403": {
                        "description": "Missing knowledge_base.view permission"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send the token returned by /api/v1/auth/register or /api/v1/auth/login as 'Authorization: Bearer {token}'.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}